[LayoutNG] Detect forced breaks even when out of space.
Forced breaks need to be reported even if we're out of space, so check
for forced breaks before we check whether we're out of space.

Currently, being recognized as forced or not only makes a difference for
column balancing, but it will also make a difference for margins at
fragmentainer breaks, once we have implemented that correctly.

Cq-Include-Trybots: master.tryserver.chromium.linux:linux_layout_tests_layout_ng
Change-Id: I7db3820124b25e91b3997e43fae17e7718047f80